fetch-agent/requirements.txt
uagents
uagents-core
httpx
python-dotenv
